requirements.txt
aiogram==3.26.0
APScheduler==3.11.2
asyncpg==0.31.0
colorlog==6.10.1
colorama==0.4.6
google-generativeai==0.8.6
pyrofork==2.3.69
TgCrypto==1.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
SQLAlchemy==2.0.48
sqlmodel==0.0.37
